tests/wpt/tests/imagebitmap-renderingcontext/context-creation-offscreen.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Canvas's ImageBitmapRenderingContext test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
 <link rel="help" href="https://html.spec.whatwg.org/multipage/canvas.html#the-imagebitmap-rendering-context">
<script>
test(function() {
    var width = 10;
    var height = 10;
    var canvas = new OffscreenCanvas(width, height);
    var ctx = canvas.getContext('bitmaprenderer');
    assert_true(ctx instanceof ImageBitmapRenderingContext);
    assert_true("canvas" in ctx);
    assert_equals(canvas, ctx.canvas);
}, "Test that canvas.getContext('bitmaprenderer') returns an instance of ImageBitmapRenderingContext and ctx.canvas on a ImageBitmapRenderingContext returns the original OffscreenCanvas");

</script>
